refactor: InventoryJpaEntity 위치 변경
