+++ practices/inventory/src/main/java/com/grizz/inventoryapp/inventory/repository/jpa/InventoryJpaRepository.java
@@ -1,6 +1,6 @@
 package com.grizz.inventoryapp.inventory.repository.jpa;
 
-import com.grizz.inventoryapp.inventory.repository.entity.InventoryEntity;
+import com.grizz.inventoryapp.inventory.repository.jpa.entity.InventoryEntity;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
